tmp fix circular import again and remove extra newline
